converted HLLinStep to LinStep
