+++ algocert/solvers/sdp_cgal_solver/__init__.py
@@ -1,11 +1,11 @@
+from algocert.basic_algorithm_steps.basic_linear_step import BasicLinearStep
 from algocert.basic_algorithm_steps.block_step import BlockStep
-from algocert.basic_algorithm_steps.linear_step import LinearStep
 from algocert.basic_algorithm_steps.max_with_vec_step import MaxWithVecStep
 from algocert.basic_algorithm_steps.min_with_vec_step import MinWithVecStep
 from algocert.basic_algorithm_steps.nonneg_orthant_proj_step import \
     NonNegProjStep
 from algocert.high_level_alg_steps.box_proj_step import BoxProjStep
-from algocert.high_level_alg_steps.hl_linear_step import HighLevelLinearStep
+from algocert.high_level_alg_steps.linear_step import LinearStep
 from algocert.high_level_alg_steps.nonneg_lin_step import NonNegLinStep
 from algocert.init_set.box_set import BoxSet
 from algocert.init_set.centered_l2_ball_set import CenteredL2BallSet
@@ -16,15 +16,15 @@ from algocert.solvers.sdp_cgal_solver.set_canonicalizers.box_set import \
     box_set_canon
 from algocert.solvers.sdp_cgal_solver.set_canonicalizers.centered_l2_ball_set import \
     centered_l2_ball_canon
-from algocert.solvers.sdp_cgal_solver.step_canonicalizers.hl_linear_step import \
-    hl_linear_step_canon
+from algocert.solvers.sdp_cgal_solver.step_canonicalizers.linear_step import \
+    linear_step_canon
 from algocert.solvers.sdp_cgal_solver.step_canonicalizers.nonneg_lin_step import \
     nonneg_lin_canon
 from algocert.solvers.sdp_cgal_solver.step_canonicalizers.nonneg_orthant_proj_step import \
     nonneg_orthant_proj_canon
 
 HL_TO_BASIC_STEP_METHODS = {
-    HighLevelLinearStep: hl_linear_step_canon,
+    LinearStep: linear_step_canon,
 }
 
 OBJ_CANON_METHODS = {
@@ -37,7 +37,7 @@ SET_CANON_METHODS = {
 }
 
 STEP_CANON_METHODS = {
-    HighLevelLinearStep: hl_linear_step_canon,
+    LinearStep: linear_step_canon,
     NonNegLinStep: nonneg_lin_canon,
     NonNegProjStep: nonneg_orthant_proj_canon,
 }
