+++ algocert/solvers/sdp_solver/__init__.py
@@ -1,11 +1,11 @@
+from algocert.basic_algorithm_steps.basic_linear_step import BasicLinearStep
 from algocert.basic_algorithm_steps.block_step import BlockStep
-from algocert.basic_algorithm_steps.linear_step import LinearStep
 from algocert.basic_algorithm_steps.max_with_vec_step import MaxWithVecStep
 from algocert.basic_algorithm_steps.min_with_vec_step import MinWithVecStep
 from algocert.basic_algorithm_steps.nonneg_orthant_proj_step import \
     NonNegProjStep
 from algocert.high_level_alg_steps.box_proj_step import BoxProjStep
-from algocert.high_level_alg_steps.hl_linear_step import HighLevelLinearStep
+from algocert.high_level_alg_steps.linear_step import LinearStep
 from algocert.high_level_alg_steps.nonneg_lin_step import NonNegLinStep
 from algocert.init_set.box_set import BoxSet
 from algocert.init_set.centered_l2_ball_set import CenteredL2BallSet
@@ -34,14 +34,14 @@ from algocert.solvers.sdp_solver.set_canonicalizers.linf_ball_set import \
     linf_ball_canon
 from algocert.solvers.sdp_solver.set_canonicalizers.vec_span_set import \
     vec_span_canon
+from algocert.solvers.sdp_solver.step_canonicalizers.basic_linear_step import (
+    basic_linear_step_bound_canon, basic_linear_step_canon)
 from algocert.solvers.sdp_solver.step_canonicalizers.block_step import (
     block_step_bound_canon, block_step_canon)
 from algocert.solvers.sdp_solver.step_canonicalizers.box_proj_step import \
     box_proj_step_canon
-from algocert.solvers.sdp_solver.step_canonicalizers.hl_linear_step import \
-    hl_linear_step_canon
-from algocert.solvers.sdp_solver.step_canonicalizers.linear_step import (
-    linear_step_bound_canon, linear_step_canon)
+from algocert.solvers.sdp_solver.step_canonicalizers.linear_step import \
+    linear_step_canon
 from algocert.solvers.sdp_solver.step_canonicalizers.max_with_vec_step import (
     max_vec_bound_canon, max_vec_canon)
 from algocert.solvers.sdp_solver.step_canonicalizers.min_with_vec_step import (
@@ -62,7 +62,7 @@ SET_CANON_METHODS = {
 
 STEP_CANON_METHODS = {
     BlockStep: block_step_canon,
-    LinearStep: linear_step_canon,
+    BasicLinearStep: basic_linear_step_canon,
     NonNegLinStep: nonneg_lin_canon,
     NonNegProjStep: nonneg_orthant_proj_canon,
     MaxWithVecStep: max_vec_canon,
@@ -76,7 +76,7 @@ RLT_CANON_SET_METHODS = {
 
 RLT_CANON_STEP_METHODS = {
     BlockStep: block_step_bound_canon,
-    LinearStep: linear_step_bound_canon,
+    BasicLinearStep: basic_linear_step_bound_canon,
     NonNegProjStep: nonneg_orthant_proj_bound_canon,
     MaxWithVecStep: max_vec_bound_canon,
     MinWithVecStep: min_vec_bound_canon,
@@ -89,6 +89,6 @@ OBJ_CANON_METHODS = {
 }
 
 HL_TO_BASIC_STEP_METHODS = {
-    HighLevelLinearStep: hl_linear_step_canon,
+    LinearStep: linear_step_canon,
     BoxProjStep: box_proj_step_canon,
 }
